requirements.txt
opencv-python
ultralytics
mediapipe
flask
gTTS
google-genai
tqdm
scikit-learn